Fixing double prefix issue of files in the root directory
Hey Petr, thanks for the module! Very useful.

I found one small issue with it. Less files in the root directory get prefixed twice. This is probably due to some changes in the 'glob' module. Glob function is called twice (lines 42 and 43) in the bootstrap-prefixer.js file - once for files in root, and then for all .less files in the root directory and all 1-level subdirectories. The first call is not necessary anymore, as the second one 'catches' the root and all subdirectories.